+++ Renderer/premake5.lua
@@ -1,0 +1,19 @@
+project "Renderer"
+    kind "StaticLib"
+    language "C++"
+    targetdir "bin/%{cfg.buildcfg}"
+
+    files {
+        "src/**.h",
+        "src/**.cpp"
+    }
+
+    filter "configurations:Debug"
+        defines { "DEBUG" }
+        symbols "On"
+
+    filter "configurations:Release"
+        defines { "NDEBUG" }
+        optimize "On"
+
+filter {}
