+++ Renderer/src/vec3.cpp
@@ -1,0 +1,11 @@
+#include "vec3.h"
+
+Renderer::Vec3::Vec3()
+    : x(0.0f), y(0.0f), z(0.0f)
+{
+}
+
+Renderer::Vec3::Vec3(float x, float y, float z)
+    : x(x), y(y), z(z)
+{
+}
