+++ Sandbox/premake5.lua
@@ -1,0 +1,21 @@
+project "Sandbox"
+    kind "ConsoleApp"
+    language "C++"
+    targetdir "bin/%{cfg.buildcfg}"
+
+    files {
+        "src/**.h",
+        "src/**.cpp"
+    }
+
+    links { "Renderer" }
+
+    filter "configurations:Debug"
+        defines { "DEBUG" }
+        symbols "On"
+
+    filter "configurations:Release"
+        defines { "NDEBUG" }
+        optimize "On"
+
+filter {}
