+++ premake5.lua
@@ -1,17 +1,5 @@
-workspace "HelloWorld"
+workspace "Renderer-cpp"
    configurations { "Debug", "Release" }
 
-project "HelloWorld"
-   kind "ConsoleApp"
-   language "C++"
-   targetdir "bin/%{cfg.buildcfg}"
-
-   files { "src/**.h", "src/**.cpp" }
-
-   filter "configurations:Debug"
-      defines { "DEBUG" }
-      symbols "On"
-
-   filter "configurations:Release"
-      defines { "NDEBUG" }
-      optimize "On"
+include "Sandbox"
+include "Renderer"
