--- src/main.cpp
@@ -1,7 +1,0 @@
-#include <iostream>
-
-int main(void) {
-    std::cout << "Hello, World!" << std::endl;
-    return 0;
-}
-
